style: #6 회원가입 화면 스타일 분리
