+++ src/components/signup/EmailCheckButton.tsx
@@ -1,31 +1,17 @@
-import React from 'react';
-import {TouchableOpacity, StyleSheet, Platform, Text} from 'react-native';
+import React, {Component} from "react";
+import {TouchableOpacity, Text} from 'react-native';
+import styles from '../../styles/signup/EmailCheckButton';
 
-const EmailCheckButton = () => {
-    const styles = StyleSheet.create({
-        buttonStyle: {
-            marginLeft: 10,
-            padding: 14,
-            backgroundColor: '#2BC63B',
-            borderRadius: 8
-        },
-
-        buttonTitle: {
-            marginTop: 4,
-            textAlign: 'center',
-            color: 'white', 
-            fontSize: 15,
-            fontFamily: 'GmarketSansTTFMedium'
-        }
-    });
-
-    return (
-        <TouchableOpacity style = {styles.buttonStyle}>
-            <Text style = {styles.buttonTitle}>
-                인증하기
-            </Text>
-        </TouchableOpacity>
-    );
+class EmailCheckButton extends Component {
+    render () {
+        return (
+            <TouchableOpacity style = {styles.buttonStyle}>
+                <Text style = {styles.buttonTitle}>
+                    인증하기
+                </Text>
+            </TouchableOpacity>
+        );
+    };
 };
 
 export default EmailCheckButton;
