+++ src/styles/signup/DomainInputbox.js
@@ -1,0 +1,13 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginTop: 4, 
+        padding: 8,
+        backgroundColor: 'white',
+        borderColor: '#FFFFF',
+        borderRadius: 8
+    }
+});
+
+export default styles;
