+++ src/styles/signup/EmailCheckButton.js
@@ -1,0 +1,20 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    buttonStyle: {
+        marginLeft: 10,
+        padding: 14,
+        backgroundColor: '#2BC63B',
+        borderRadius: 8
+    },
+
+    buttonTitle: {
+        marginTop: 4,
+        textAlign: 'center',
+        color: 'white', 
+        fontSize: 15,
+        fontFamily: 'GmarketSansTTFMedium'
+    }
+});
+
+export default styles;
