+++ src/styles/signup/IDInputbox.js
@@ -1,0 +1,19 @@
+import {StyleSheet, Platform} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginTop: 4, 
+        padding: 8,
+        backgroundColor: 'white',
+        borderColor: '#FFFFF',
+        borderRadius: 8
+    },
+
+    inputText: {
+        color: 'gray', 
+        fontSize: 15,
+        fontFamily: 'GmarketSansTTFLight'
+    }
+});
+
+export default styles;
