+++ src/styles/signup/NameInputbox.js
@@ -1,0 +1,20 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginTop: 4, 
+        marginRight: 20,
+        padding: 8,
+        backgroundColor: 'white',
+        borderColor: '#FFFFF',
+        borderRadius: 8
+    },
+
+    inputText: {
+        color: 'gray', 
+        fontSize: 15,
+        fontFamily: 'GmarketSansTTFLight'
+    }
+});
+
+export default styles;
