+++ src/styles/signup/SignUpButton.js
@@ -1,0 +1,21 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginTop: 30,
+        marginLeft: 3,
+        marginRight: 3,
+        padding: 15,
+        backgroundColor: '#2BC63B',
+        borderRadius: 8
+    },
+
+    inputText: {
+        color: 'white',
+        textAlign: 'center',
+        fontSize: 20,
+        fontFamily: 'GmarketSansTTFMedium'
+    }
+});
+
+export default styles;
